Extended factor nodes

Extended factor nodes

Index

ForneyLab.@composite

Composite

ForneyLab.@composite — Macro.

The @composite macro allows for defining custom (composite) nodes. Composite nodes allow for implementating of custom update rules that may be computationally more efficient or convenient. A composite node can be defined with or without an internal model. For detailed usage instructions we refer to the composite_nodes demo.

source