Extended rules

Extended rules

Index

ForneyLab.@expectationPropagationRule

ForneyLab.@marginalRule

ForneyLab.@naiveVariationalRule

ForneyLab.@structuredVariationalRule

ForneyLab.@sumProductRule

Expectation propagation rule

ForneyLab.@expectationPropagationRule — Macro.

@expectationPropagationRule registers a expectation propagation update rule by defining the rule type and the corresponding methods for the outboundType and isApplicable functions. If no name (type) for the new rule is passed, a unique name (type) will be generated. Returns the rule type.

source

Marginal rule

ForneyLab.@marginalRule — Macro.

@marginalRule registers a marginal update rule for a (joint) marginal by defining the rule type and the corresponding methods for the isApplicable functions. If no name (type) for the new rule is passed, a unique name (type) will be generated. Returns the rule type.

source

Naive variational rule

ForneyLab.@naiveVariationalRule — Macro.

@naiveVariationalRule registers a variational update rule for the naive (mean-field) factorization by defining the rule type and the corresponding methods for the outboundType and isApplicable functions. If no name (type) for the new rule is passed, a unique name (type) will be generated. Returns the rule type.

source

Structured variational rule

ForneyLab.@structuredVariationalRule — Macro.

@structuredVariationalRule registers a variational update rule for the structured factorization by defining the rule type and the corresponding methods for the outboundType and isApplicable functions. If no name (type) for the new rule is passed, a unique name (type) will be generated. Returns the rule type.

source

Sum-product rule

ForneyLab.@sumProductRule — Macro.

@sumProductRule registers a sum-product update rule by defining the rule type and the corresponding methods for the outboundType and isApplicable functions. If no name (type) for the new rule is passed, a unique name (type) will be generated. Returns the rule type.

source