Graph (low-level)

Graph (low-level)

Index

ForneyLab.Edge

ForneyLab.Interface

ForneyLab.Terminal

Edge

ForneyLab.Edge — Type.

An Edge joins two interfaces (half-edges) a and b.

source

Interface

ForneyLab.Interface — Type.

An Interface belongs to a FactorNode and represents a half-edge. An Interface has at most one partner interface, with wich it forms an edge.

source

Terminal

ForneyLab.Terminal — Type.

Description:

Terminal is a special type of node that is only used in the internal
graph of a CompositeNode. A Terminal is used to terminate an Edge in the
internal graph that is linked to an interface of the CompositeNode.

A Terminal is linked to an interface of the
CompositeNode containing the Terminal.

Interfaces:

1. out

Construction:

Terminal(id=:some_id)

source