Scheduler (low-level)

Scheduler (low-level)

Index

ForneyLab.MarginalRule

ForneyLab.MarginalScheduleEntry

ForneyLab.MarginalUpdateRule

ForneyLab.MessageUpdateRule

ForneyLab.RecognitionFactor

ForneyLab.ScheduleEntry

Marginal rule

ForneyLab.MarginalRule — Type.

MarginalRule{factortype} specifies a joint marginal update rule with respect to a node of type `factortype`.

source

Marginal schedule entry

ForneyLab.MarginalScheduleEntry — Type.

A MarginalScheduleEntry defines a marginal computation. The marginal_update_rule <: MarginalUpdateRule defines the rule that is used to calculate the (joint) marginal over target.

source

Marginal update rule

ForneyLab.MarginalUpdateRule — Type.

A MarginalUpdateRule specifies how a (joint) marginal is calculated from incoming messages (and a node function).

source

Message update rule

ForneyLab.MessageUpdateRule — Type.

A MessageUpdateRule specifies how a Message is calculated from the node function and the incoming messages. Use subtypes(MessageUpdateRule) to list the available rules.

source

Recognition factor

ForneyLab.RecognitionFactor — Type.

A RecognitionFactor specifies the subset of variables that comprise a joint factor in the recognition factorization.

source

Schedule entry

ForneyLab.ScheduleEntry — Type.

A ScheduleEntry defines a message computation. The msg_update_rule <: MessageUpdateRule defines the rule that is used to calculate the message coming out of interface.

source