Construct algorithm code

Construct algorithm code

Index

ForneyLab.expectationPropagationAlgorithm

ForneyLab.freeEnergyAlgorithm

ForneyLab.sumProductAlgorithm

ForneyLab.variationalAlgorithm

ForneyLab.variationalExpectationPropagationAlgorithm

Expectation propagation algorithm

ForneyLab.expectationPropagationAlgorithm — Function.

Create a sum-product algorithm to infer marginals over variables, and compile it to Julia code

source

Free energy algorithm

ForneyLab.freeEnergyAlgorithm — Function.

The freeEnergyAlgorithm function accepts a RecognitionFactorization and returns (if possible) Julia code for computing the variational free energy with respect to the argument recognition factorization and corresponding FactorGraph (model).

source

Sum-product algorithm

ForneyLab.sumProductAlgorithm — Function.

Create a sum-product algorithm to infer marginals over variables, and compile it to Julia code

source

Variational algorithm

ForneyLab.variationalAlgorithm — Function.

Create a variational algorithm to infer marginals over a recognition distribution, and compile it to Julia code

source

Variational expectation propagation algorithm

ForneyLab.variationalExpectationPropagationAlgorithm — Function.

Create a variational EP algorithm to infer marginals over a recognition distribution, and compile it to Julia code

source