Execute an algorithm

Execute an algorithm

Index

ForneyLab.Message

ForneyLab.PointMass

ForneyLab.ProbabilityDistribution

Message

ForneyLab.Message — Type.

Encodes a message, which is a probability distribution with a scaling factor

source

Point mass

ForneyLab.PointMass — Type.

PointMass is an abstract type used to describe point mass distributions. It never occurs in a FactorGraph, but it is used as a probability distribution type.

source

Probability distribution

ForneyLab.ProbabilityDistribution — Type.

Encodes a probability distribution as a FactorNode of type family with fixed interfaces

source