Helpers

Helpers

Index

ForneyLab.@symmetrical

ForneyLab.ensureMatrix

ForneyLab.isApproxEqual

ForneyLab.isRoundedPosDef

ForneyLab.leaftypes

ForneyLab.mat

ForneyLab.trigammaInverse

Description

ForneyLab.mat — Method.

Helper function to construct 1x1 Matrix

source

ForneyLab.@symmetrical — Macro.

Duplicate a method definition with the order of the first two arguments swapped. This macro is used to duplicate methods that are symmetrical in their first two input arguments, but require explicit definitions for the different argument orders. Example:

@symmetrical function prod!(x, y, z)
    ...
end

source

ForneyLab.ensureMatrix — Method.

ensureMatrix: cast input to a Matrix if necessary

source

ForneyLab.isApproxEqual — Method.

isApproxEqual: check approximate equality

source

ForneyLab.isRoundedPosDef — Method.

isRoundedPosDef: is input matrix positive definite? Round to prevent fp precision problems that isposdef() suffers from.

source

ForneyLab.leaftypes — Method.

leaftypes(datatype) returns all subtypes of datatype that are leafs in the type tree.

source

ForneyLab.trigammaInverse — Method.

trigammaInverse(x): solve trigamma(y) = x for y.

Uses Newton's method on the convex function 1/trigramma(y). Iterations converge monotonically. Based on trigammaInverse implementation in R package "limma" by Gordon Smyth: https://github.com/Bioconductor-mirror/limma/blob/master/R/fitFDist.R

source