Schedule an algorithm

Schedule an algorithm

Index

ForneyLab.MarginalSchedule

ForneyLab.RecognitionFactorization

ForneyLab.Schedule

ForneyLab.currentRecognitionFactorization

ForneyLab.expectationPropagationSchedule

ForneyLab.sumProductSchedule

ForneyLab.variationalExpectationPropagationSchedule

ForneyLab.variationalSchedule

Marginal Schedule

ForneyLab.MarginalSchedule — Type.

A MarginalSchedule defines the update order for marginal computations.

source

Recognition Factorization

ForneyLab.RecognitionFactorization — Type.

A RecognitionFactorization holds a collection of (non-overlapping) recognition factors that specify the recognition factorization over a factor graph that is used for variational inference.

source

Schedule

ForneyLab.Schedule — Type.

A Schedule defines the update order for message computations.

source

Сurrent recognition factorization

ForneyLab.currentRecognitionFactorization — Function.

Return currently active RecognitionFactorization. Create one if there is none.

source

Expectation propagation schedule

ForneyLab.expectationPropagationSchedule — Function.

expectationPropagationSchedule() generates a expectation propagation message passing schedule.

source

Sum-Product schedule

ForneyLab.sumProductSchedule — Function.

sumProductSchedule() generates a sum-product message passing schedule that computes the marginals for each of the argument variables.

source

Variational expectation propagation schedule

ForneyLab.variationalExpectationPropagationSchedule — Function.

variationalExpectationPropagationSchedule() generates an expectation propagation message passing schedule that is limited to the recognition_factor. Updates on EP sites are computed with an ExpectationPropagationRule.

source

Variational schedule

ForneyLab.variationalSchedule — Function.

variationalSchedule() generates a variational message passing schedule that computes the marginals for each of the recognition distributions in the recognition factor.

source