Actors API

How to create a custom Actor

At first custom actor should implement a custom method for the as_actor function. Rocket.jl also provides a number of helper actor abstract types with predefined as_actor method behavior (see Traits API section). Optionally custom actor may implement a custom method for the is_exhausted function which is used to check if actor can handle any further message events.

using Rocket

struct MyCustomActor end

as_actor(::Type{<:MyCustomActor}) = Rocket.BaseActorTrait{Int}()

is_exhausted(actor::MyCustomActor) = # some custom logic here, false by default for all actors.

or

using Rocket

struct MyCustomActor <: Actor{Int} end # Automatically specifies BaseActorTrait{Int} behavior.

Additionally custom actor must provide a custom methods for on_next!, on_error! and/or on_complete! functions. Depending on specified actor trait behavior some methods may or may not be optional.

using Rocket

struct MyCustomActor <: Actor{Int} end

Rocket.on_next!(actor::MyCustomActor, data::Int)  = # custom logic here
Rocket.on_error!(actor::MyCustomActor, err)       = # custom logic here
Rocket.on_complete!(actor::MyCustomActor)         = # custom logic here

or

using Rocket

struct MyCustomCompletionActor <: CompletionActor{Int} end

Rocket.on_complete!(actor::MyCustomCompletionActor) = # custom logic here

Traits

Rocket.ActorTrait — Type

Abstract type for all possible actor traits

See also: Actor, BaseActorTrait, NextActorTrait, ErrorActorTrait, CompletionActorTrait, InvalidActorTrait

Rocket.as_actor — Function

as_actor(::Type)

This function checks actor trait behavior specification. May be used explicitly to specify actor trait behavior for any object.

See also: ActorTrait

Rocket.BaseActorTrait — Type

Base actor trait specifies actor to listen for all next!, error! and complete! events.

See also: ActorTrait, Actor

Rocket.NextActorTrait — Type

Next actor trait specifies actor to listen for next! events only.

See also: ActorTrait, NextActor

Rocket.ErrorActorTrait — Type

Error actor trait specifies actor to listen for error! events only.

See also: ActorTrait, ErrorActor

Rocket.CompletionActorTrait — Type

Completion actor trait specifies actor to listen for complete! events only.

See also: ActorTrait, CompletionActor

Rocket.InvalidActorTrait — Type

Default actor trait behavior for any object. Actor with such a trait specificaion cannot be used as a valid actor in subscribe! function. Doing so will raise an error.

Types

Rocket.AbstractActor — Type

Abstract type for any actor object

See also: Actor, NextActor, ErrorActor, CompletionActor

Rocket.Actor — Type

Can be used as a super type for common actor. Automatically specifies a BaseActorTrait trait behavior. Each Actor must implement its own methods for on_next!(actor, data), on_error!(actor, err) and on_complete!(actor) functions.

See also: AbstractActor, BaseActorTrait, ActorTrait, on_next!, on_error!, on_complete!

Rocket.NextActor — Type

Can be used as a super type for "next-only" actor. Automatically specifies a NextActorTrait trait behavior. Each NextActor must implement its own methods for on_next!(actor, data) function only.

See also: AbstractActor, NextActorTrait, ActorTrait, on_next!

Rocket.ErrorActor — Type

Can be used as a super type for "error-only" actor. Automatically specifies a ErrorActorTrait trait behavior. Each ErrorActor must implement its own methods for on_error!(actor, err) function only.

See also: AbstractActor, ErrorActorTrait, ActorTrait, on_error!

Rocket.CompletionActor — Type

Can be used as a super type for "completion-only" actor. Automatically specifies a CompletionActorTrait trait behavior. Each CompletionActor must implement its own methods for on_complete!(actor) function only.

See also: AbstractActor, CompletionActorTrait, ActorTrait, on_complete!

Events

Rocket.next! — Function

next!(actor, data)

This function is used to deliver a "next" event to an actor with some data

See also: AbstractActor, on_next!

Rocket.error! — Function

error!(actor, err)

This function is used to deliver a "error" event to an actor with some err

See also: AbstractActor, on_error!

Rocket.complete! — Function

complete!(actor)

This function is used to deliver a "complete" event to an actor

See also: AbstractActor, on_complete!

Rocket.is_exhausted — Function

is_exhausted(actor)

This function is used to check if actor can handle any further message events

See also: AbstractActor

Rocket.on_next! — Function

on_next!(actor, data)

Both Actor and NextActor objects must implement its own method for on_next! function which will be called on "next" event.

See also: Actor, NextActor

Rocket.on_error! — Function

on_error!(actor, err)

Both Actor and ErrorActor objects must implement its own method for on_error! function which will be called on "error" event.

See also: Actor, ErrorActor

Rocket.on_complete! — Function

on_complete!(actor)

Both Actor and CompletionActor objects must implement its own method for on_complete! function which will be called on "complete" event.

See also: Actor, ErrorActor

Factory

Rocket.AbstractActorFactory — Type

Abstract type for all possible actor factories

See also: Actor

Rocket.create_actor — Function

create_actor(::Type{L}, factory::F) where L where { F <: AbstractActorFactory }

Actor creator function for a given factory F. Should be implemented explicitly for any AbstractActorFactory object

See also: AbstractActorFactory, MissingCreateActorFactoryImplementationError

Rocket.MissingCreateActorFactoryImplementationError — Type

This error will be throw if Julia cannot find specific method of 'create_actor()' function for given actor factory

See also: AbstractActorFactory, create_actor

Errors

Rocket.InvalidActorTraitUsageError — Type

This error will be thrown if next!, error! or complete! functions are called with invalid actor object

See also: next!, error!, complete!, InvalidActorTrait

Rocket.InconsistentSourceActorDataTypesError — Type

This error will be thrown if next! function is called with inconsistent data type

See also: AbstractActor, Subscribable, next!

Rocket.MissingDataArgumentInNextCall — Type

This error will be thrown if next! function is called without data argument

See also: next!

Rocket.MissingErrorArgumentInErrorCall — Type

This error will be thrown if error! function is called without err argument

See also: error!

Rocket.ExtraArgumentInCompleteCall — Type

This error will be thrown if complete! function is called with extra data/err argument

See also: complete!

Rocket.MissingOnNextImplementationError — Type

This error will be thrown if Julia cannot find specific method of 'on_next!()' function for given actor and data

See also: on_next!

Rocket.MissingOnErrorImplementationError — Type

This error will be thrown if Julia cannot find specific method of 'on_error!()' function for given actor

See also: on_error!

Rocket.MissingOnCompleteImplementationError — Type

This error will be thrown if Julia cannot find specific method of 'on_complete!()' function for given actor and data

See also: on_next!