Subjects API

Traits

Rocket.SubjectTrait — Type

Abstract type for all possible subject traits

See also: ValidSubject, InvalidSubject, as_subject

Rocket.as_subject — Function

as_subject(::Type)

This function checks subject trait behavior specification. Should be used explicitly to specify subject trait behavior for any object type.

See also: SubjectTrait

Rocket.ValidSubject — Type

Valid subject trait behavior

See also: SubjectTrait

Rocket.InvalidSubject — Type

Default subject trait behavior for all types.

See also: SubjectTrait

Factory

Rocket.AbstractSubjectFactory — Type

Abstract type for all possible subject factories

See also: SubjectTrait, ValidSubject, InvalidSubject

Rocket.create_subject — Function

create_subject(::Type{L}, factory::F) where L where { F <: AbstractSubjectFactory }

Actor creator function for a given factory F. Should be implemented explicitly for any AbstractActorFactory object

See also: AbstractSubjectFactory, MissingCreateActorFactoryImplementationError

Rocket.MissingCreateSubjectFactoryImplementationError — Type

This error will be throw if Julia cannot find specific method of 'create_subject()' function for given subject factory

See also: AbstractSubjectFactory, create_subject

Errors

Rocket.InvalidSubjectTraitUsageError — Type

InvalidSubject usage error

See also: as_subject

Rocket.InconsistentSubjectDataTypesError — Type

InconsistentSubjectDataTypesError

See also: as_subject