File Observable

Rx.file — Function

file(path::String)

Creation operator for the SyncFileObservable with a given path.

See also: SyncFileObservable

Rx.SyncFileObservable — Type

SyncFileObservable(path::String)

File observable, which synchronously emits content of the file line by line as a String objects on subscription.

See also: file, Subscribable