Scan Operator

Rocket.scan — Function

scan(::Type{R}, scanFn::Function, seed::Union{R, Nothing} = nothing) where R

Creates a scan operator, which applies a given accumulator scanFn function to each value emmited by the source Observable, and returns each intermediate result with an optional seed value. If a seed value is specified, then that value will be used as the initial value for the accumulator. If no seed value is specified, the first item of the source is used as the seed.

Arguments

::Type{R}: the type of data of transformed value, may be or may not be the same as type of input source

scanFn::Function: accumulator function with (data::T, current::R) -> R signature

seed::R: optional initial value for accumulator function

Producing

Stream of type <: Subscribable{R}

Examples

using Rocket

source = from([ 1, 2, 3 ])
subscribe!(source |> scan(Vector{Int}, (d, c) -> [ c..., d ], Int[]), logger())
;

# output

[LogActor] Data: [1]
[LogActor] Data: [1, 2]
[LogActor] Data: [1, 2, 3]
[LogActor] Completed

See also: AbstractOperator, RightTypedOperator, ProxyObservable, reduce, logger

Description

Combines all values emitted by the source, using an accumulator function that joins a new source value with the past accumulation. This is similar to reduce, but emits the intermediate accumulations.

Returns an Observable that applies a specified accumulator function to each item emitted by the source Observable. If a seed value is specified, then that value will be used as the initial value for the accumulator. If no seed value is specified, the first item of the source is used as the seed.

Performance tip

For performance reasons, do not use lambda based operators in production Julia code. Either use macro helpers to generate efficient versions of operators (like @CreateMapOperator() and/or @CreateFilterOperator(), etc..) or implement your own operators without using lambda functions.

@CreateScanOperator macro

For performance reasons, the Rocket.jl library provides a special macro for creating custom (and pure) scan operators.

Rocket.@CreateScanOperator — Macro

@CreateScanOperator(name, L, R, scanFn)

Creates a custom scan operator, which can be used as nameScanOperator().

Arguments

name: custom operator name

L: type of input data

R: type of output data after scanFn projection

scanFn: accumulator function, assumed to be pure

Generates

nameScanOperator(seed::R) function

Producing

Stream of type <: Subscribable{R}

Examples

using Rocket

@CreateScanOperator(IntoArray, Int, Vector{Int}, (d, c) -> [ c..., d ])

source = from([ 1, 2, 3 ])
subscribe!(source |> IntoArrayScanOperator(Int[]), logger())
;

# output

[LogActor] Data: [1]
[LogActor] Data: [1, 2]
[LogActor] Data: [1, 2, 3]
[LogActor] Completed

See also: AbstractOperator, TypedOperator, ProxyObservable, scan, logger

See also

Operators